Assorted perf and memory optimizations driven by profiling
Reduces inverse bloom filter size 10x to reduce overhead in writing it
to Clickhouse.

Also moves duplicate filter test into batch construction thread so it
only tests and adds events when they're about to be added to a batch,
rather then when they've been processed. This avoids the possibility of
a batch duplicate filter including events that have been processed but
not written on restart if the processing threads get ahead of the batch
write thread.