Handle replication row column mismatch and add skip anonymized fields config
Now skips rows where the current schema columns don't match the binlog
event columns. This is a scenario that was causing an error because
there's no good way to map binlog column data that's out of sync with the
current state of the database.

Also adds a skip anonymized fields config to explicitly exclude fields
that might match one of the anonymized fields arguments. This is needed
because Golang's regex engine doesn't support negative lookaheads for
performance reasons.